+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The following changes have been completed, and will be included in the next tagg
 
 ### Fix
 
+- color conversions from colors with alpha channel of `0` no longer end up with an alpha channel of `1`
 - `makeCssXxxColorFromCssColor()` and `makeCssXxxColorFromConversionModel()` constructors now do the necessary color-space conversions internally
   - added `convertConversionModelToOklchColorSpace()` helper
   - added `convertConversionModelToSrgbColorSpace()` helper